requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
python-slugify==8.0.4
pytz==2024.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
